requirements.txt
argcomplete
argh
pyyaml
networkx
tornado